+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/sign/Signer.kt
@@ -100,12 +100,18 @@ interface Signer {
     interface ECDSA : AlgTrait {
         override val signatureAlgorithm: SignatureAlgorithm.ECDSA
         override val publicKey: CryptoPublicKey.EC
+
+        @SecretExposure
+        override fun exportPrivateKey(): KmmResult<CryptoPrivateKey.EC.WithPublicKey>
     }
 
     /** A [Signer] that signs using RSA. */
     interface RSA : AlgTrait {
         override val signatureAlgorithm: SignatureAlgorithm.RSA
         override val publicKey: CryptoPublicKey.RSA
+
+        @SecretExposure
+        override fun exportPrivateKey(): KmmResult<CryptoPrivateKey.RSA>
     }
 
     /** Some [Signer]s are retrieved from a signing provider, such as a key store, and have a string [alias]. */
@@ -128,6 +134,9 @@ interface Signer {
     suspend fun sign(data: SignatureInput): SignatureResult<*>
     suspend fun sign(data: ByteArray) = sign(SignatureInput(data))
     suspend fun sign(data: Sequence<ByteArray>) = sign(SignatureInput(data))
+
+    /** Performs (EC)DH key agreement. Might ask for user confirmation first if this [Signer] [mayRequireUserUnlock]. */
+    suspend fun keyAgreement(publicKey: CryptoPublicKey): KmmResult<ByteArray>
 
     companion object {
         fun Ephemeral(configure: DSLConfigureFn<EphemeralSigningKeyConfiguration> = null) =