+++ supreme/src/iosMain/kotlin/at/asitplus/signum/supreme/sign/PrivateKeySignerImpl.kt
@@ -1,9 +1,11 @@
 package at.asitplus.signum.supreme.sign
 
 import at.asitplus.KmmResult
+import at.asitplus.catching
 import at.asitplus.signum.indispensable.*
 import at.asitplus.signum.internals.*
 import at.asitplus.signum.supreme.*
+import at.asitplus.signum.supreme.agreement.performKeyAgreement
 import kotlinx.cinterop.ExperimentalForeignApi
 import platform.Foundation.NSData
 import platform.Security.SecKeyCreateSignature
@@ -48,6 +50,13 @@ protected constructor(
         }
     }
 
+    @OptIn(ExperimentalForeignApi::class)
+    override suspend fun keyAgreement(publicKey: CryptoPublicKey) = catching {
+        if (this !is Signer.ECDSA)
+            throw UnsupportedCryptoException("iOS does not support non-EC Diffie-Hellman.")
+        performKeyAgreement(secKey, publicKey)
+    }
+
     @SecretExposure
     override fun exportPrivateKey() = KmmResult.success(privateKey)
 }
@@ -58,11 +67,21 @@ class ECPrivateKeySigner(
     override val privateKey: CryptoPrivateKey.EC.WithPublicKey,
     override val signatureAlgorithm: SignatureAlgorithm.ECDSA,
     override val publicKey: CryptoPublicKey.EC
-) : PrivateKeySigner(privateKey.toSecKey().getOrThrow(), signatureAlgorithm), Signer.ECDSA
+) : PrivateKeySigner(privateKey.toSecKey().getOrThrow(), signatureAlgorithm), Signer.ECDSA {
+
+    @SecretExposure
+    override fun exportPrivateKey() =
+        super.exportPrivateKey().mapCatching { it as CryptoPrivateKey.EC.WithPublicKey }
+}
 
 @OptIn(ExperimentalForeignApi::class)
 class RSAPrivateKeySigner(
     override val privateKey: CryptoPrivateKey.RSA,
     override val signatureAlgorithm: SignatureAlgorithm.RSA,
     override val publicKey: CryptoPublicKey.RSA
-) : PrivateKeySigner(privateKey.toSecKey().getOrThrow(), signatureAlgorithm), Signer.RSA
+) : PrivateKeySigner(privateKey.toSecKey().getOrThrow(), signatureAlgorithm), Signer.RSA {
+
+    @SecretExposure
+    override fun exportPrivateKey() =
+        super.exportPrivateKey().mapCatching { it as CryptoPrivateKey.RSA }
+}