+++ supreme/src/iosMain/kotlin/at/asitplus/signum/supreme/sign/VerifierImpl.kt
@@ -9,6 +9,8 @@ import at.asitplus.signum.internals.*
 import at.asitplus.signum.indispensable.iosEncoded
 import at.asitplus.signum.indispensable.nativeDigest
 import at.asitplus.signum.indispensable.secKeyAlgorithmPreHashed
+import at.asitplus.signum.supreme.HazardousMaterials
+import at.asitplus.signum.supreme.toSecKey
 import at.asitplus.signum.supreme.dsl.DSL
 import at.asitplus.signum.supreme.UnsupportedCryptoException
 import kotlinx.cinterop.ExperimentalForeignApi
@@ -47,16 +49,8 @@ internal actual fun checkAlgorithmKeyCombinationSupportedByRSAPlatformVerifier
 {
 }
 
-private fun MemScope.toSecKey(key: CryptoPublicKey): SecKeyRef =
-    corecall {
-        SecKeyCreateWithData(key.iosEncoded.toNSData().giveToCF(), cfDictionaryOf(
-            kSecAttrKeyClass to kSecAttrKeyClassPublic,
-            kSecAttrKeyType to when (key) {
-                is CryptoPublicKey.EC -> kSecAttrKeyTypeEC
-                is CryptoPublicKey.RSA -> kSecAttrKeyTypeRSA
-            }), error)
-    }.also { defer { CFRelease(it) }}
 
+@OptIn(HazardousMaterials::class)
 private fun verifyImpl(signatureAlgorithm: SignatureAlgorithm, publicKey: CryptoPublicKey,
                        data: SignatureInput, signature: CryptoSignature,
                        config: PlatformVerifierConfiguration) {