+++ supreme/src/jvmMain/kotlin/at/asitplus/signum/supreme/sign/EphemeralKeysImpl.kt
@@ -2,22 +2,19 @@ package at.asitplus.signum.supreme.sign
 
 import at.asitplus.KmmResult
 import at.asitplus.catching
-import at.asitplus.signum.indispensable.CryptoPrivateKey
-import at.asitplus.signum.indispensable.CryptoPublicKey
-import at.asitplus.signum.indispensable.CryptoSignature
-import at.asitplus.signum.indispensable.SignatureAlgorithm
-import at.asitplus.signum.indispensable.fromJcaPublicKey
-import at.asitplus.signum.indispensable.getJCASignatureInstance
-import at.asitplus.signum.indispensable.getJCASignatureInstancePreHashed
-import at.asitplus.signum.indispensable.jcaName
-import at.asitplus.signum.indispensable.parseFromJca
+import at.asitplus.signum.indispensable.*
+import at.asitplus.signum.supreme.HazardousMaterials
 import at.asitplus.signum.supreme.SecretExposure
+import at.asitplus.signum.supreme.hazmat.jcaPrivateKey
 import at.asitplus.signum.supreme.signCatching
 import com.ionspin.kotlin.bignum.integer.base63.toJavaBigInteger
 import java.security.KeyPairGenerator
 import java.security.PrivateKey
+import java.security.interfaces.ECPrivateKey
+import java.security.interfaces.RSAPrivateKey
 import java.security.spec.ECGenParameterSpec
 import java.security.spec.RSAKeyGenParameterSpec
+import javax.crypto.KeyAgreement
 
 
 @SecretExposure
@@ -54,17 +51,26 @@ sealed class EphemeralSigner (internal val privateKey: PrivateKey, private val p
         }
     }
 
-    @SecretExposure
-    override fun exportPrivateKey(): KmmResult<CryptoPrivateKey.WithPublicKey<*>> = catching { CryptoPrivateKey.decodeFromDer(
-        privateKey.encoded) as CryptoPrivateKey.WithPublicKey<*> }
-
     protected abstract fun parseFromJca(bytes: ByteArray): CryptoSignature.RawByteEncodable
+
+    final override suspend fun keyAgreement(publicKey: CryptoPublicKey) = catching {
+        KeyAgreement.getInstance(when (this) {
+            is EC -> "ECDH"
+            is RSA -> "DH"
+        }).also {
+            it.init(privateKey)
+            it.doPhase(publicKey.toJcaPublicKey().getOrThrow(), true)
+        }.generateSecret()
+    }
 
     open class EC internal constructor (config: JvmEphemeralSignerCompatibleConfiguration, privateKey: PrivateKey,
               override val publicKey: CryptoPublicKey.EC, override val signatureAlgorithm: SignatureAlgorithm.ECDSA)
         : EphemeralSigner(privateKey, config.provider), Signer.ECDSA {
 
         override fun parseFromJca(bytes: ByteArray) = CryptoSignature.EC.parseFromJca(bytes).withCurve(publicKey.curve)
+
+        @SecretExposure
+        final override fun exportPrivateKey() = CryptoPrivateKey.fromJcaPrivateKey(privateKey as ECPrivateKey)
     }
 
     open class RSA internal constructor (config: JvmEphemeralSignerCompatibleConfiguration, privateKey: PrivateKey,
@@ -72,6 +78,9 @@ sealed class EphemeralSigner (internal val privateKey: PrivateKey, private val p
         : EphemeralSigner(privateKey, config.provider), Signer.RSA {
 
         override fun parseFromJca(bytes: ByteArray) = CryptoSignature.RSAorHMAC.parseFromJca(bytes)
+
+        @SecretExposure
+        final override fun exportPrivateKey() = CryptoPrivateKey.fromJcaPrivateKey(privateKey as RSAPrivateKey)
     }
 }